fix: 修复重启后已启动的服务状态不对的问题

+++ jarboot-core/src/main/java/com/mz/jarboot/core/utils/LogUtils.java
@@ -4,6 +4,8 @@ import ch.qos.logback.classic.Logger;
 import ch.qos.logback.classic.LoggerContext;
 import ch.qos.logback.classic.encoder.PatternLayoutEncoder;
 import ch.qos.logback.classic.spi.ILoggingEvent;
+import ch.qos.logback.core.Appender;
+import ch.qos.logback.core.ConsoleAppender;
 import ch.qos.logback.core.FileAppender;
 import com.mz.jarboot.api.constant.CommonConst;
 import com.mz.jarboot.core.constant.CoreConstant;
@@ -21,7 +23,7 @@ public class LogUtils {
     private static Logger logger;
     private static String logDir;
 
-    public static void init(String home, String server, String sid) {
+    public static void init(String home, String server, String sid, boolean persist) {
         if (null != logger) {
             return;
         }
@@ -32,31 +34,41 @@ public class LogUtils {
                 "[%file:%line] %msg%n");
         ple.setContext(lc);
         ple.start();
-        FileAppender<ILoggingEvent> fileAppender = new FileAppender<>();
-        StringBuilder sb = new StringBuilder();
-        sb
-                .append(home)
-                .append(File.separator)
-                .append("logs")
-                .append(File.separator)
-                .append(server);
-        logDir = sb.toString();
-        sb.append(File.separator)
-                .append("jarboot-")
-                .append(server)
-                .append('-')
-                .append(sid)
-                .append(".log");
-        String log =sb.toString();
-        fileAppender.setFile(log);
-        fileAppender.setEncoder(ple);
-        fileAppender.setContext(lc);
-        fileAppender.start();
+
+        Appender<ILoggingEvent> appender;
+        if (persist) {
+            FileAppender<ILoggingEvent> fileAppender = new FileAppender<>();
+            StringBuilder sb = new StringBuilder();
+            sb
+                    .append(home)
+                    .append(File.separator)
+                    .append("logs")
+                    .append(File.separator)
+                    .append(server);
+            logDir = sb.toString();
+            sb.append(File.separator)
+                    .append("jarboot-")
+                    .append(server)
+                    .append('-')
+                    .append(sid)
+                    .append(".log");
+            String log =sb.toString();
+            fileAppender.setFile(log);
+            fileAppender.setEncoder(ple);
+            fileAppender.setContext(lc);
+            fileAppender.start();
+            appender = fileAppender;
+        } else {
+            ConsoleAppender<ILoggingEvent> consoleAppender = new ConsoleAppender<>();
+            consoleAppender.setEncoder(ple);
+            consoleAppender.setContext(lc);
+            consoleAppender.start();
+            appender = consoleAppender;
+        }
 
         //模块中所有日志均使用该名字获取
         logger = (Logger) LoggerFactory.getLogger(CoreConstant.LOG_NAME);
-        logger.addAppender(fileAppender);
-        writePidFile(sid);
+        logger.addAppender(appender);
     }
 
     public static Logger getLogger() {
@@ -85,6 +97,9 @@ public class LogUtils {
 
     public static void deletePidFile(String sid) {
         File pid = FileUtils.getFile(logDir, sid + CommonConst.PID_EXT);
+        if (!pid.exists()) {
+            return;
+        }
         try {
             FileUtils.forceDelete(pid);
         } catch (Exception exception) {
